LayoutTests/imported/w3c/web-platform-tests/speculation-rules/speculation-tags/cross-site-prefetch.https-expected.txt


FAIL cross-site prefetch should not have Sec-Speculation-Tags be sent assert_equals: expected "__no_tags__" but got "tag1"
